requirements.txt
replicate>=0.26.0,<1.0
python-dotenv>=1.0,<2.0
PyYAML>=6.0,<7.0
litellm>=1.35.0,<2.0

